+++ lib/shared/src/mentions/query.ts
@@ -1,7 +1,7 @@
 import type { RangeData } from '../common/range'
 import {
-    type ContextMentionProvider,
     type ContextMentionProviderID,
+    type ContextMentionProviderMetadata,
     FILE_CONTEXT_MENTION_PROVIDER,
     SYMBOL_CONTEXT_MENTION_PROVIDER,
 } from './api'
@@ -17,8 +17,7 @@ export interface MentionQuery {
     provider: ContextMentionProviderID | null
 
     /**
-     * The user's text input, to be interpreted as a fuzzy-matched query. It is stripped of any
-     * prefix characters that indicate the {@link MentionQuery.provider}, such as `#` for symbols.
+     * The user's text input, to be interpreted as a fuzzy-matched query.
      */
     text: string
 
@@ -45,20 +44,19 @@ export interface MentionQuery {
  */
 export function parseMentionQuery(
     query: string,
-    contextMentionProviders: Pick<ContextMentionProvider, 'id' | 'triggerPrefixes'>[]
+    provider: Pick<ContextMentionProviderMetadata, 'id'> | null
 ): MentionQuery {
+    if (provider) {
+        return { provider: provider.id, text: query }
+    }
+
     if (query === '') {
         return { provider: null, text: '' }
     }
 
+    // Special-case '#' as a trigger prefix for symbols.
     if (query.startsWith('#')) {
         return { provider: SYMBOL_CONTEXT_MENTION_PROVIDER.id, text: query.slice(1) }
-    }
-
-    for (const provider of contextMentionProviders) {
-        if (provider.triggerPrefixes.some(trigger => query.startsWith(trigger))) {
-            return { provider: provider.id, text: query }
-        }
     }
 
     const { textWithoutRange, maybeHasRangeSuffix, range } = extractRangeFromFileMention(query)